refactor: TableService 메서드 정리
