+++ src/test/java/kitchenpos/application/integration/ApplicationIntegrationTest.java
@@ -1,9 +1,9 @@
 package kitchenpos.application.integration;
 
-import kitchenpos.application.ProductService;
 import kitchenpos.application.mengroup.MenuGroupService;
 import kitchenpos.application.menu.MenuService;
 import kitchenpos.application.order.OrderService;
+import kitchenpos.application.product.ProductService;
 import kitchenpos.application.table.TableGroupService;
 import kitchenpos.application.table.TableService;
 import kitchenpos.domain.common.Money;
